Add files via upload
The clue that was found yesterday morning was crap. This is a working function we can use. The only thing we may want to consider is adding a check to force capitalization if there are no capital letters in the final result. In my testing I always got a string with at least 1 upper case, but the code doesn't technically check it, so it theoretically could happen.